format archives in client
