+++ src/_data/js.js
@@ -1,9 +1,9 @@
-/** JS web component data */
+/** JS file data */
 const fg = require('fast-glob');
 
 module.exports = async function () {
   return {
-    filePaths: (await fg('src/components/**/*.js')).map((path) =>
+    filePaths: (await fg('src/js/**/*.js')).map((path) =>
       path.replace('src/', './')
     ),
   };
