+++ src/_includes/header.njk
@@ -1,5 +1,5 @@
 <wrg-header></wrg-header>
 
 <script type="module-shim">
-  import('./components/wrg-header.js');
+  import('./js/wrg-header.js');
 </script>
