+++ src/archive.njk
@@ -12,5 +12,5 @@ eleventyExcludeFromCollections: true
 </div>
 
 <script type="module-shim">
-  import('./components/wrg-replay.js');
+  import('./js/wrg-replay.js');
 </script>
