+++ src/archives.njk
@@ -19,6 +19,6 @@ layout: layout.njk
 </section>
 
 <script type="module-shim">
-  import('./components/wrg-archives-count.js');
-  import('./components/wrg-index.js');
+  import('./js/wrg-archives-count.js');
+  import('./js/wrg-index.js');
 </script>
