+++ src/index.njk
@@ -20,6 +20,6 @@ layout: layout.njk
 </section>
 
 <script type="module-shim">
-  import('./components/wrg-archives-count.js');
-  import('./components/wrg-search.js');
+  import('./js/wrg-archives-count.js');
+  import('./js/wrg-search.js');
 </script>
