+++ src/js/wrg-archives-count.js
@@ -1,5 +1,5 @@
 import { html, css, LitElement } from 'lit';
-import config from '../wrg-config.json';
+import config from '../wrg-runtime-config.json';
 
 customElements.define(
   'wrg-archives-count',
